Update SUSYTools tau trigger SF implementation
The tau trigger group has released R21 SFs :-)
The existing SUSYTools implementation has just been updated to fix a crash due to the change in the naming convention (!31241).
However, the list of supported triggers needs to be extended (basically, two new sets of triggers have been introduced in 2018, they are named "medium1_tracktwoEF" and "mediumRNN_tracktwoMVA", which are more modern/efficient versions than "medium1_tracktwo").

This MR (ATLSUSYBGF-520) adds all triggers supported by the tau trigger group.
It also fixes a couple of things, like the HLT/offline trigger matching that has to use dR=0.2, instead of the default 0.1.
The name of the trigger used in the matching was also missing "HLT_".

To retrieve the proper SF in GetTauTriggerEfficiencySF, one has to pass the name of the calibration file, e.g. HLT_tau25_medium1_tracktwoEF_OR_mediumRNN_tracktwoMVA, rather than the name of the trigger leg.

I hope in the near future, there will be a more user-friendly implementation that will compute a per-event SF, instead of the per-leg SF we currently provide.
But for that we'll need trigger efficiencies in addition to trigger SFs, which is not yet provided by the tau trigger group.
Ideally, this would go into TrigGlobalEfficiencyCorrectionTool. Ideally...

Cheers,
Bertrand