Fixed rotation angle for MDT BMG chambers
The rotation angle and shift of MDT BMG chambers were not properly
printed because the center of BMG chambers are not aligned with
the sector center in XY-plane. This patch corrects for that
and return a rotation angle to the next sector if the angular separation
of the chamber and the next sector center is less than half of the
sector open angle.