Removed the generation of pileup event weights for DAOD_PHYSLITE.
This will require analysers to use the PRW tool to generate their
"correct" pileup weights when running over the DAOD_PHYSLITE files.
But it will simplify the PHYSLITE production a whole lot, as none
of the tools running in the DAOD_PHYSLITE production actually need
the event weights. They just need the random run numbers to be
generated.