Refactor some of the GX2F material code
This commit refactors some of the code that is responsible for adding
material to tracks in the Trk::GlobalChi2Fitter. It's still a complete
mess, but this eases the problem a little bit, I suppose. We split a
large 600-odd line function into a few smaller ones, improving
readability. Prelude for further changes to this code.