Re-added a search for FastJetContrib in Rivet_i.
While this code, and even Rivet, does not publicly depend on fjcontrib,
Rivet (the generator) does privately depend on it. So for the correct
link- and run-time environment it's healthy to keep looking for
FastJetContrib.

But note that the package is not using ${FASTJETCONTRIB_INCLUDE_DIRS}
or ${FASTJETCONTRIB_LIBRARIES} afterwards.