tauRecTools: remove member variables in TauPi0ScoreCalculator
To calculate the BDT score, we need to provide a list of input
variables. Previously, the input variables are declared as memeber
variables, which make the class stateful. These member variables
are now replaced with local variables.

Add a new class, i.e. BDTHelper, to help calculate the BDT score.
This class is similar with TRTBDT in HelperFunction. In the future,
we will remove TRTBDT.