tauRecTools: replace TRTBDT with BDTHelper in TauJetBDTEvaluator
Previously, we use TRTBDT to help calculate the BDT score. Since TRTBDT
is not thread safe, it is replaced with BDTHelper now. In addtion, an
obsolete property, i.e. isGrad, is removed.

Change the return type of getGradBoostMVA from StatusCode to float. If
there is any problem when calculating the score, -999 will be returned.