Updates for DerivationFrameworkAnalysisTests
Two updates here. First, keeping the dcube output and enabling the
web-display of the plots. That should be the final step to being able to
have "physics validation" of the TRUTH derivations.

Second is adding a check after the art download step, in order to make
sure that the inputs are actually available before the job launches.