Changes to add new RegSelTools for the Calorimeter tables
The RegSelectrorHashMap has a large number of changes simply to make the
class callable from const methods of classs with RegSelectorHAshMap data
members. This is not a full rewriting of the class, so much of the class
has been left as it is. At some point we may well want to replace this
class completely.

In addition, the new RegSelCaloTool class has been added along with the
RegSelTool since the lookup table class is different. Templating the class
might have been better, but in the end, all the actual embeded lookup table
operations are different, and the class is essentially just a wrapper to
these calls. We could have abstracted the lookup table calls, into further
lightweight wrappers but that is something that can be left for another
day.

There is also a new RegSelCondAlg_LAr class which actually creates the
lookup tables. Sadly, this included an unwanted dependence on the pixel
cabling to determine the EventIDRange for the conditions data. If there
is any Calo Conditions data for the cabling, this could be used, but for
the time being we use the pixel cabling. One there is full Calo cabling
this dependence will be removed. This is a new class, but essentially
contains a cut and paste of the code in LArRegionSelectorTable, with
some tidying up of the layout and renaming of some variables. The actual
code itself might benefit from being rewritten, but this is beyond the
scope of this merge request, which is simply to add the new tools.
A full rewrite should be up to the Calo developers.

There are new maker function for the RegSelTools in RegSelToolConfig.py