Minor tau adjustment for DAOD_PHYS (truthJetLink)
Hello,

This MR is adding the element link between reconstructed taus and truth jets.
"truthJetLink" is created centrally (by TauTruthCommon) when making derivations, but as it was not part of the tau smart slimming list, it was missing from DAOD_PHYS.
This link can be recomputed at analysis level by re-applying tau truth matching, but for convenience, it has been added to the tau smart slimming.
Also, the link is now built against AntiKt4TruthDressedWZJets in DAOD_PHYS (and only in this format), instead of AntiKt4TruthJets used by default.
Ref: ATLASG-1502.

Cheers,
Bertrand