RDBAccessSvc: Thread-safety check updates
Pending changes to the thread-safety checker warn about potentially
unsafe uses of coral::AttributeList.  Add annotations here to suppress
these warnings as needed.
    
As part of this, need to change getTagDetails so that it is not returning
an AttributeList by value.  Instead, pass a list via nonconst reference.
    
Also move the thread-safety checker flag file to the header directory.