AthenaMonitoringKernel: Fix string filling with generator
Fix and unit testing for the following use-case:
```
Monitored::Scalar<std::string>( "DetID", [&](){return "SCT";} );
```
Filling from a generated string was not supported and instead of a
compiler error it actually made the code crash because it tried to
initialize a string from a double (e.g. `std::string m_value(0)`).