IOVSvc: Fixing some IOVSvc memory leaks.
Fix a couple memory leaks in IOVSvc::createCondObj.

The address returned from getRangeFromDB was leaked.
Fixed by changing this method to return the address via a unique_ptr.

An object is read via StoreGateSvc::createObj.  This returns a DataObject*,
which is a DataBucket holding the actual payload object.
The payload object is added to the conditions container, but the
DataBucket is leaked.  Fixed to make sure the DataBucket is freed.