Implement photon radiation enhancement for photon+jet process in PowhegControl
This functionality requires further updates from the Powheg authors to work. The
Powheg authors have already given us details of how it should work in the future,
so the changes here are partially based on that. But testing is needed before we
know if the enhancement works, in particular in conjunction with getting
multiple QCD scale and PDF weights via reweighting.