AthenaMonitoringKernel: Extend Rebinable class to 2D
Allow the creation of "rebinable" 2D histograms. Implementation-wise
this is handled by converting the existing `HistogramFillerRebinable1D`
class into a "Mixin" class that is templated with the base class (1D or
2D) and the histogram axis that is supposed to be rebinable.

Typedefs are provided to handle the existing 1D case and the three new
cases to rebin 2D histograms along x, y and both axes. An additional
unit test suite for the 2D case has been added.