TrigOpMonitor: Book IOVDbSvc histogram unconditionally.
The IOVDbSvc histograms were booked conditionally on the existence of
the IOVDbSvc. In very simple jobs (e.g. HelloWorld) this could lead to
different behavior between the first and second run during stop/start as
the IOVDbSvc may have been created (but in generally not used) after the
histogram booking took place in the first run.