Updates to MGC to get gridpack generation error-free
This updates a couple lines in the infrastructure and several of the
tests to be able to make a gridpack without getting an error from the
transform. The last remaining part was to touch an output txt file and
use that to drive the final step of the transform job.