Protect against empty RDO containers in muon decoding
Adding some protection for cases where we have empty RDO containers in the RDO to PRD decoding for the CSCs and MDTs (the RPCs and TGCs appear to have such protection already). The RDO containers can be empty when running the decoding for the trigger where we don't decode the full detector.