Adding debug file output to log in MadGraphControl
This should solve a long-standing issue in MadGraphControl, that there
is a debug file generated by MadGraph that doesn't appear in the log,
but can be extremely useful (or even necessary) for debugging. Adding a
hook in the error checking function that looks for such a file and
prints it into the log in the case that it is found. Seems to be
behaving based on very early tests.