Allowing HTML errors in MGC and fixing gridpack madspin setup
Identified an error scenario when making a gridpack that results in an
error to do with HTML files moving around. Obviously we don't care about
the HTML files to begin with, so I've reported the error to the
developers but am allowing it as an exception for the time being.

Additionally, I discovered with our tests that the run name for a
gridpack run is hard-coded inside of MadGraph, even if you tell it to
change the run name. This is an unfortunate feature, but it basically
means that we have to account for two possible run names: the one that
we constantly and consistently tell it to use, and this other one that
it has hard-coded. The code has been updated to cope with that
situation, I hope in a reasonable way.