MadGraphControl adding an error and Higgs parameter settings
Adding an error to the known list, another stty error that comes because
of the handling of input and output between the python and fortran
layers in MG5_aMC. Hope to investigate these with the authors soon.

Also splitting the top parameters function into settings for SM
parameters and for top parameters (calling the SM parameter settings
from the top parameter settings, so that their behavior is unchanged).

Also splitting off the Higgs parameter settings, and adding the setting
of the Higgs decay fractions as a part of a helper function. Calling
this from the top parameter setting function as well. Likely future
setups will want to call this top parameter setting function.