Two tweaks to MGC to help with params and for SUSY
First change to the pre-include just sets default masses for the heavy
SM-sector Higgs bosons to be super heavy (decoupled)

Second is a tweak to the param card handling. There are cases in param
cards (particularly for mixing matrices) where the key is really
several numbers, and the value is only the last. This improves handling
for those cases, both by making sure all the keys are included and by
allowing weird spacing in the searches (so for example '1 1' will match
'1   1'). This should make life quite a bit easier.