Implementation of some Chris commets (Thanks!)
The main changes involve a modification in the options in the .config
file to add pT and Mass upper limits and to add the TruthLabelAccessor
name (BoostJetTaggers).

The code will not work with the old config files so they have to be
updated in the CalibArea. The new name of the TruthLabelAccessor is not
implemented in Athena yet so two different verisons of the new config
files will be sent to Chris.