DAOD_PHYS: apply extremely loose thinning cut on RNN tau ID
Hello,

This MR adds an extremely loose thinning cut on the (flattened) RNN tau identification score in DAOD_PHYS and PHYSLITE formats (ATLASG-1502).
The score is required to be greater than 0.01, which is 99% efficient for true taus, but rejects significant jet background already.
This is much looser than the loosest tau ID working point (very loose = cut at 0.05, for a target efficiency of 95%).
The motivation for this cut is to reduce a bit further the DAOD size, by removing reconstructed taus that will never be used for analysis.
The lower cut of 0.01 is the loosest working point that will be supported by the Fake Tau Task Force, for the fake tau estimate.

Size estimates from ART tests:
* DAOD_PHYS data before modif:
```
     Disk Size         Fraction    Category Name
       0.885 kb        0.032       tau
       2.098 kb        0.076       InDet
      27.507 kb        1.000       Total
```
* DAOD_PHYS data after modif:
```
       0.578 kb        0.021       tau
       1.897 kb        0.070       InDet
      27.007 kb        1.000       Total
```
* DAOD_PHYS MC before modif:
```
       1.232 kb        0.027       tau
       2.841 kb        0.062       InDet
      45.787 kb        1.000       Total
```
* DAOD_PHYS MC after modif:
```
       0.708 kb        0.016       tau
       2.482 kb        0.055       InDet
      44.804 kb        1.000       Total
```

Modest gain, 0.5 kb/event for data, 1kb/event for MC, but always good to take.

Cheers,
Bertrand