Trying to tidy MGC gridpack and cluster setup
First, fixing a generic error in the code, that if ATHENA_PROC_NUMBER is
set to zero (but is set), the code would have a great deal of trouble.
Now recognizing such a setting as serial.

Second, trying to clean up and justify the usage of bin/generate_events,
transferring all arguments through the cards rather than on the command
line (including the setting of number of cores and so on). This seems to
get batch system running up again for LO and NLO (new!).