Patch to MP for generation jobs
During generation jobs, there are no input files and the maxEvents
setting is often -1 (that's how we're told to test, normally). This
results in a condition that the transform recognizes as having fewer
input events than processes, so athenaMP is disabled. Adding an explicit
catch for this condition to try to allow MP in these job configurations.
Local tests seem to be working.