InDetTestPixelLayer: Remove slot-specific caching.
InDetTestPixelLayer had been caching information in members.

This was dealt with in the MT migration by making the caching slot-specific,
introducing significant complexity.

However, none of the cached information is actually used in current Athena!

The methods getFracGood(), getExtrapolPixelEta(), and getExtrapolPixelPhi()
are never called.  Just remove them.

The residuals methods are also never called.  Since these are a bit more 
involved, don't remove them entirely.  But if we really wanted to use them,
it would be better to return the information that is being cached
to the client (perhaps in a cookie) rather then keeping it in the tool.
So, retain these methods but remove the caching, along with a note
that the interface should probably be changed if this functionality
is seriously used again.