silence warning about missing ToolsRegistry tool
This should fix https://its.cern.ch/jira/browse/ATLASRECTS-5413

I really did just the minimal thing to silence this warning, but if
you ask me this seems not like a happy/healthy design.  I'm already
far from convinced (to put it mildly) that it is a good idea for
analysis code to talk to the ToolStore directly, but for code running
in reconstruction to talk to the ASG ToolStore to figure out if a tool
exists and if not to create it backhandedly without ever interacting
with the ToolSvc at all, well that seems less than ideal, or at the
very least somewhat unusual.