Prioritizing gridpack output in madspin
Before, we just took the first hit. This looks a little more carefully
at the output, to say that if we had a hit that looks like it came from
a grid pack it should get used first, and otherwise we take the first
hit (same default). It protects against the case that the gridpack has
some old events stashed in it.