Replace deprecated PRD association tool by PRDtoTrackMap.
The map is needed for to create decorations for the combined track.
This maps is already created during reconstruction but also has to be
recorded to StoreGate to be available by the decorator. In case the
decorations are applied starting from an ESD the PRD association map
has to be recreated. The latter is needed for IDTRKVALID with
dumpTrtInfo enabled.

Also configure the TRT_ToT_dEdx tool which is needed by the TSOS decorator
rather then relying on this tool being configured elsewhere in the
job.