Implement python access to trigger configuration
Provides access to the L1 and HLT menus and prescale-sets and to the HLT
job options. Missing is the bunchgroup access, although all the
infrastructure exists

For each type of configuration one access class exists, which can be
called with a file name or at trigger alias and db-key

The class holds an order-preserving dictionary of the configuration
data, which can be accessed directly. Convenience functions for commonly
used data exist.

An executable has been added to make use of this functionality