Consolidate muon trigger hypotheses
Since the TrigMuonEFMSonlyHypo alg/tool was basically a copy of the TrigMuonEFCombinerHypo alg/tool, have combined them into a single hypo alg/tool (TrigMuonEFHypo). The only difference was what type of muons were being considered (MSonly vs Combined), so this is now a configurable property of the tool. The corresponding configuration has been updated to configure the different instances of the hypo alg/tool and the monitoring configuration has similarly been condensed.