Central PDF setup for SUSY MadGraph
This adds one of the central PDF fragments to the common SUSY setup. It
was discussed in the Background Forum meeting on 14.4 and agreed that
this would be done. This brings the SUSY setup in line with the current
PMG recommendations, and has the nice benefit of including systematic
variations in a standard way.