Removing spurious line in PHYS configuration
This line has two problems: 1) it shows up only in the MC section, when
this container is in fact for data and MC; 2) it seems to corrupt the
output of the derivation framework, so that the interface is not
available for the container. Removing the line still gets the container
in the output file (since it is included already in the smart slimming)
and fixes the interface. Solves a comment from ATLASG-1502