PyUtils: cling bug workaround.
When we run aclic, root produces a DSO and tries to link it.
As past of this, ROOT tries to find all the libraries that the new
DSO depends on by listing all its symbols and then walking through
all libraries on LD_LIBRARY_PATH to find a definition.
Due to a bug, cling will crash if it encounters a separate-debug
library in doing so.  Pretty much every library contains
the gprof stub __gmon_start__ as a hidden weak symbol.
Cling ends up searching for this symbol, and since it is hidden,
it will end up trying all libraries.  In Athena builds, it eventually
hits a debug library and dies.

Work around by injecting a dummy defintion of __gmon_start__.

Fixes unit test failure in master.