BPHY8: Update of options and reduction of sets of isolation and closest track variables kept
This update of BPHY8 in preparation of the revalidation run on data15
and data16 includes the following modifications:
- Update settings for muon calibration and smearing tool to latest
  recommendations.
- Reduction of the number of produced isolation and closest track
  variable combinations, including thinning decorations for combinations
  not needed for the analysis, considerably reducing the size of the
  derivation output.
- Thinning the ID track collection turned on (again).
- Fix for properly keeping the closest track per candidate in the
  thinned ID track collection.

The ID track thinning and the reduction/thinning of isolation and
closest track variables kept reduces the DAODs size by a factor
of approx. 3.5 w.r.t. the last BPHY8 version which had only been
intended for validation studies on a few selected runs.