Fix bug: 2D binned variable evaluation point finder always returns last bin
The if condition of the check if the evaluation point is in the current bin always
resolves to "False" as a value can never be both greater and smaller than one bin
edge. Correct the check so that it checks if the evaluation point is between the
lower edge of the current and the lower edge of the next bin by adding a +1 to the
bin iterator in the second condition.