Make T->P and P->T methods for HepMcParticleLinkCnv_p1 consistent
In the 21.0 branch an eventIndex of zero means that the transient `HepMcParticleLink` looks
at the first event in the `McEventCollection`, otherwise the value returned by
`HepMcParticleLink::eventIndex()` represents the GenEvent::event_number of a `GenEvent` in
the `McEventCollection` and this is used to find the appropriate `GenEvent` in the `McEventCollection`
when using the link.

NB There is a weakness here that if more than one `GenEvent` has the same event number all links will
point to the first `GenEvent` which matches. This will be fixed after the migration to `HepMC3`.

Also dropped references to separate pile-up `McEventCollections` as the use of these is not supported
in 21.0 releases.