Extend HepMcParticleLinkCnv unit tests to check links to background GenEvents
This commit extends `HepMcParticleLinkCnv_p1_test.cxx` and `HepMcParticleLinkCnv_p2_test.cxx`
to check links to `GenParticle` objects which are not part of the first `GenEvent` in the
`McEventCollection`. It also adds a check to ensure that links are correct when the position
of a `GenEvent` in the `McEventCollection` rather than its event number is used.