Special handling for evgen MP
In evgen we need to handle MP a bit differently. This is my attempt,
which seems to avoid errors in the transform itself. The problem is that
we normally disable MP during the job, so that the output file appears
as just a simple output file with whatever name has been assigned by the
transform. There aren't a bunch of workers. Very happy to have input on
whether there's a better way to do this!