AthenaKernel: Fix for extending mixed ranges
When we extend the run+lbn of a mixed range, we were updating the range
key in the top-level run+lbn map, but we were not changing the keys
in the dependent tsmap.  That means a client retrieving the condition
could see an old range for the entry.  Fix by updating the run+lbn
stop value for all entries in the dependent tsmap.

See ATLASRECTS-5467.