Attempt to fix issue seen in ATLPHYSVAL-630
After a detailed investigation detailed in ATLPHYSVAL-630, it was determined that since release/21.0.41 particles
were reset to their original properties each time they were passed back from ISF to Geant4, rather than keeping
track of the energy lost in the propagation through the detector so far. The fix does not seem to alter the output
for the FullG4 simulator in ISF or `AtlasG4_tf.py` in local tests, but this should be carefully checked with higher
statistics.