Add low-pt ditau to DAOD_PHYSVAL
Hello,

This MR adds the low-pt ditau reconstruction to DAOD_PHYSVAL.
It is scheduled in the AOD->DAOD step.
The "DiTauJetsLowPt" collection has been added to "AllVariables" (the smart slimming list is anyway very close to all variables).
That adds 0.8 kb/event to the output, on top of the current 271 kb, for the data ART file.
Note to ourselves: we should start monitoring the ditau distributions in PHYSVAL...

Cheers,
Bertrand