Add xAODEventInfoAthenaPool package to AthGeneration and AthSimulation projects
Currently the `xAODEventInfoCnvTest` fails in the AthGeneration and AthSimulation nightlies.
This is because these projects do not properly support reading/writing `xAOD::EventAuxInfo_v2`.
Adding the `Event/xAOD/xAODEventInfoAthenaPool` package to these projects should fix this.
And will be required anyway for when we eventually switch simulation and generation jobs to
writing `xAOD::EventInfo` rather than `EventInfo`. Should fix ATLASSIM-4525