Robustness improvements for HepMcParticleLinkCnv_p2.cxx and related TP converters
In the case that the standard production workflow is being used
then the first event in the `McEventCollection` is the only one
where the `eventIndex` could exceed the maximum value of unsigned
short. In this case we can work around the issue by using an
`m_mcEvtIndex` of zero as a special case, in which `m_mcEvtIndex`
should be interpreted as the position of the `GenEvent` in the
`McEventCollection` rather than the value of `GenEvent::event_number()`.

Similar changes are made to SiHitCollectionCnv_p3.cxx and TRT_HitCollectionCnv_p4.cxx.