Update transient HepMcParticleLink to use unsigned int rather than unsigned short to store the eventIndex
This change should allow the HepMcParticleLink to accomodate all allowed values of `GenEvent::event_number`
i.e. up to `INT_MAX` = 21^31 -1 = 2147483647.

A new persistent class is not being added at this time, but existing workarounds allow event numbers larger
than max unsigned short to be correctly recovered for links pointing to the first event in the McEventCollection.
In the standard production workflow, pile-up background event numbers are reset to to numbers in the range
0 to 2^15 -1, which should be OK up to <mu> values of 840 with current settings.

A new persistent version of HepMcParticleLink will be introduced after the migration to HepMC3.