TrigSteeringEventTPCnv: Implement MT thinning.
Previously, HLTResult objects were thinned either using the slimming
functionality of ThinningSvc or by modifying it in-place using
an algorithm.  Neither of these will work for MT jobs, and
ThinningSvc is anyway being removed.

Add a hook to the TP converter to query ThinningCache for
TrigNavigation thinning and apply it if we find it.