Updating MET to reduce stored variables
The MET requires the pre-sampler fraction. This MR tries to
pre-calculate the pre-sampler fraction, store it as a decoration on the
jets, pass it along via both the baseline MET configuration and PHYSLITE
configuration, and use it if available in METMaker. Should improve the
issues seen in MET agreement with DAOD_PHYSLITE