extend and clean up the data handle interfaces/implementations
There were a number of methods that seemed meaningful to have, and a
number of others we would never be able to use in AnalysisBase, so I
enabled the former and removed the later.

There are still a number of commented out methods, mostly because I
still need to implement accessing the meta-data store.