Fixing jet accessor bug in derivations
Following !32610 there was a crash in derivation making:

```
AODtoDAOD 22:22:47 xAOD::Jet_v1 ERROR Unknown attribute (56) requested.
(Did you ask the wrong type ?)
```

This turns out to be because the accessors are more complicated than I
realized -- we need both an enum and an addition to the map. Now that's
done, so we should be good to go again. I hope the jet folks will yell
if this is going to cause any problems!