Restoring thinning to DAOD_PHYS in master
This MR restores the thinning to DAOD_PHYS, removing the lines relevant to the old thinning service, which no longer exists, dropping the thinning helper which is no longer needed, and passing to the thinning tools the stream name rather than a service. Some aspects of the thinning tools still need to be synchronised with 21.2 and so some lines are commented out. An as-yet unmigrated tool called in MuonsCommon is also commented out until the package is swept into master.