add/fix packages for AnalysisBase
The main change compared to the last time I did this is the presence
of ReadHandle/ReadHandleKey.  Also fixed up some minor issues/problems
when compiling the packages as part of AnalysisBase.

Some of the more major issues still to address is the lack of
decoration and write handles, and all the tools still using TAccept.
Also some strang problems in the constructors for the systematics data
handles.  All of those will be addressed in future merge requests.