Check validity of the RegSelCondAlg lookup table conditions data before executing
The RegSelCondAlgs should only be called if the lookup tables they create need updating.
However in serial athena, this mechanism is not implemented and so the conditions algorithms
are unaccountably called for EVERY EVENT.

As such, we need to check whether the lookup table has already been created before actually
continuing with the execute method.