Remove obsolete interfaces from FastTRT_ Digitization package
This commit removes the obsolete/redundant `ITRTFastDigitizationTool` interface and
additionally tidies up the `TRTFastDigitization` algorithms.
(During this clean-up it was noted in the `CMakeLists.txt` file that this package
has a dependency on a package which only builds a component library. This will be cleaned-up in
a separate merge request.)