Beautify the code in ByteStreamEVentInputSvc
Three things are combined into this commit: style, adoption of recent
C++ components, and some code de-duplication. The style updates
unfortunately makes this look like a full rewrite. Indentation was set
to use two spaces. The range based for loop was adopted where suitable.
Dynamically allocated memory was migrated to be managed using unique_ptr
were possible. The clearing of existing content in the Storegate was
abstrcted into a function.