Fix for ATLASRECTS-5488
In very rare cases, due to very slight variations in the functioning of different extrapolation methods, it may not be possible to create a CaloExtension for a combined track.
This is the case even though a combined track, by definition, should go through the calorimeter.
This leads to a crash in the IsolationBuilder, which requires the CaloExtension to measure the calo isolation.
So far, this has been observed exactly once, for a poor-quality MuGirl track, and as a result the check (to ensure that a CaloExtension can be built) is implemented for now only for MuGirl tracks.
In the long-term the differences in the different extrapolation methods should be investigated, but this will likely take some time.
This change also includes a cleanup of some comments and a minor fix to the MuonPrintingTool.