StoreGate: Fix WriteDecorHandle::isAvailable.
WriteDecorHandle;:isAvailable had a couple problems.  First, the isAvailable
code assumed that the object type derived from AuxVectorBase, which means
that it didn't work at all for standalone types.  Second, calling
isAvailable would populate the cached pointer without calling
the overload of typeless_dataPointer_impl.  This meant that the 
alias might not be created.

Both fixd.

Also clean up the tests so that we don't need to override the JO
search path to run them.