Fix inconsistency between PFlow and EMTopo fJVT implementation.
PFlow central jets were flagged as 'failing' the fJVT cut, requiring users to select themselves the forward jets for which to retrieve the flag.
This is now done as for EMTopo fJVT, with the jets outside the tagging range flagged as 'passing fJVT', so that users can simply retrieve the fJVT flag for all jets without any issue.