Replace FindBin calls inside inner most loop to reduce CPU time
In order to save the amount of CPU time spent inside the large number of FindBin
calls inside the inner most loop over the neighboring pixels, the FindBin calls are
replaced by the arithmetic implemented inside ROOT's FindBin function. This is possible,
because the histograms are read in from maps and do not change during the execution.
Furthermore, all histograms have equally sized bins. The necessary multiplication factor,
representing the histogram is calculated upon initialization.