Cutting away some of the outer neighboring pixels in order to reduce iterations/CPU time
In order to reduce the large number of iterations in the inner most loop
(over the nearest neighboring pixels (p,q) for every charge) I reduced the limits for the
iterators. The edge cases (p==+-2 || q==+-2) make 2/3 of the neighbors and do not change
the results significantly. We studied different high level variables to validate the small
impact. This is done without any additional computational overhead by directly changing the
calculation of the limits to be inside +-1.