Updated the xAODHIEvent package's dictionary generation.
Made sure that all potentially useful/necessary types are generated
in an organised fashion, using the helper code from xAODCore.

Removed code not necessary with modern ROOT versions.

Declared the typedef names to the generated ROOT dictionary.

Fixed the schema evolution rule for the smart pointers to the interface
containers.