Athena: Produce graphviz output during cmake configuration
Run `cmake --graphviz` by default to produce a dependency file during
the cmake configuration and install that file into the release. The
`CMakeGraphVizOptions.cmake` are tuned to get the maximum amount
of dependency information but excluding ATLAS internal targets. The
generated DOT file is ~12MB (it would be ~7MB without external targets).