Make code faster by removing unnecessary variables and technical optimizing
Remove test variables and calculations no longer needed (ramo_i_e and other *_i*
variables). Also the z dimension of the calls of FindBin is unnecessarily located
within the nn-loop and therefore evaluated too many times. When the hole has reached the
back end, ramo_h is set to zero for all iterations in nn-loop, so I added an if condition
so the necessary calculations are only performed if not the case. Same for electrons and
one can also skip the reading of the map for the final electron position when it’s true.